requirements.txt
amqp==2.5.2
aniso8601==8.0.0
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
Babel==2.8.0
billiard==3.6.2.0
celery==4.4.0
Click==7.0
Flask==1.1.1
flask-restx==0.1.1
flower==0.9.3
gunicorn==20.0.4
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
kombu==4.6.7
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.4
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
tornado==5.1.1
typed-ast==1.4.1
vine==1.3.0
Werkzeug==0.16.1
wrapt==1.11.2
zipp==2.2.0
